Creating glue catalog resource and adding set glue encryption action
