+++ c7n/resources/account.py
@@ -17,7 +17,6 @@ import json
 import time
 from botocore.exceptions import ClientError
 from datetime import datetime, timedelta
-from .aws import shape_validate
 from dateutil.parser import parse as parse_date
 from dateutil.tz import tzutc
 
@@ -1383,114 +1382,3 @@ class SetS3PublicBlock(BaseAction):
             client.put_public_access_block(
                 AccountId=r['account_id'],
                 PublicAccessBlockConfiguration=config)
-
-
-@filters.register('glue-security-config')
-class GlueEncryptionEnabled(MultiAttrFilter):
-    """Filter aws account by its glue encryption status and KMS key """
-
-    """:example:
-
-    .. yaml:
-
-      policies:
-        - name: glue-security-config
-          resource: aws.account
-          filters:
-            - type: glue-security-config
-                key: SseAwsKmsKeyId
-                value: alias/aws/glue
-
-    """
-    retry = staticmethod(QueryResourceManager.retry)
-
-    schema = {
-        'type': 'object',
-        'additionalProperties': False,
-        'properties': {
-            'type': {'enum': ['glue-security-config']},
-            'CatalogEncryptionMode': {'type': 'string'},
-            'SseAwsKmsKeyId': {'type': 'string'},
-            'ReturnConnectionPasswordEncrypted': {'type': 'boolean'},
-            'AwsKmsKeyId': {'type': 'string'}
-        }
-    }
-
-    annotation = "c7n:glue-security-config"
-    permissions = ('glue:GetDataCatalogEncryptionSettings',)
-
-    def validate(self):
-        attrs = set()
-        for key in self.data:
-            if key in ['CatalogEncryptionMode',
-                       'ReturnConnectionPasswordEncrypted',
-                       'SseAwsKmsKeyId',
-                       'AwsKmsKeyId']:
-                attrs.add(key)
-        self.multi_attrs = attrs
-        return super(GlueEncryptionEnabled, self).validate()
-
-    def get_target(self, resource):
-        if self.annotation in resource:
-            return resource[self.annotation]
-        client = local_session(self.manager.session_factory).client('glue')
-        encryption_setting = client.get_data_catalog_encryption_settings().get(
-            'DataCatalogEncryptionSettings')
-        resource[self.annotation] = encryption_setting.get('EncryptionAtRest')
-        resource[self.annotation].update(encryption_setting.get('ConnectionPasswordEncryption'))
-
-        for kmskey in self.data:
-            if not self.data[kmskey].startswith('alias'):
-                continue
-            key = resource[self.annotation].get(kmskey)
-            vfd = {'c7n:AliasName': self.data[kmskey]}
-            vf = KmsRelatedFilter(vfd, self.manager)
-            vf.RelatedIdsExpression = 'KmsKeyId'
-            vf.annotate = False
-            if not vf.process([{'KmsKeyId': key}]):
-                return []
-            resource[self.annotation][kmskey] = self.data[kmskey]
-        return resource[self.annotation]
-
-
-@actions.register('set-glue-catalog-encryption')
-class GlueDataCatalogEncryption(BaseAction):
-    """Modifies glue data catalog encryption based on specified
-    parameter
-
-    :example:
-
-    .. code-block:: yaml
-
-            policies:
-              - name: data-catalog-encryption
-                resource: account
-                filters:
-                  - type: glue-security-config
-                    CatalogEncryptionMode: DISABLED
-                actions:
-                  - type: set-glue-catalog-encryption
-                    attributes:
-                        EncryptionAtRest:
-                            CatalogEncryptionMode: SSE-KMS
-                            SseAwsKmsKeyId: alias/aws/glue
-    """
-
-    schema = type_schema(
-        'set-glue-catalog-encryption',
-        attributes={'type': 'object', "minItems": 1},
-        required=('attributes',))
-
-    permissions = ('glue:PutDataCatalogEncryptionSettings',)
-    shape = 'PutDataCatalogEncryptionSettingsRequest'
-
-    def validate(self):
-        attrs = {}
-        attrs['DataCatalogEncryptionSettings'] = self.data['attributes']
-        return shape_validate(attrs, self.shape, 'glue')
-
-    def process(self, catalog):
-        client = local_session(self.manager.session_factory).client('glue')
-        # there is one glue data catalog per account
-        client.put_data_catalog_encryption_settings(
-            DataCatalogEncryptionSettings=self.data['attributes'])
