--- tests/data/placebo/test_account_glue_connection_password_filter/kms.ListAliases_1.json
@@ -1,27 +1,0 @@
-{
-    "status_code": 200,
-    "data": {
-        "Aliases": [
-            {
-                "AliasName": "alias/skunk/trails",
-                "AliasArn": "arn:aws:kms:us-east-1:644160558196:alias/skunk/trails",
-                "TargetKeyId": "36812ccf-daaf-49b1-a24b-0eef254ffe41"
-            }
-        ],
-        "Truncated": false,
-        "ResponseMetadata": {
-            "RequestId": "d2784ff6-ab70-4e98-9210-c1b040ab5ef5",
-            "HTTPStatusCode": 200,
-            "HTTPHeaders": {
-                "x-amzn-requestid": "d2784ff6-ab70-4e98-9210-c1b040ab5ef5",
-                "cache-control": "no-cache, no-store, must-revalidate, private",
-                "expires": "0",
-                "pragma": "no-cache",
-                "date": "Tue, 23 Jul 2019 01:20:04 GMT",
-                "content-type": "application/x-amz-json-1.1",
-                "content-length": "202"
-            },
-            "RetryAttempts": 0
-        }
-    }
-}
