--- tests/data/placebo/test_account_glue_encyption_filter/glue.GetDataCatalogEncryptionSettings_1.json
@@ -1,27 +1,0 @@
-{
-    "status_code": 200,
-    "data": {
-        "DataCatalogEncryptionSettings": {
-            "EncryptionAtRest": {
-                "CatalogEncryptionMode": "SSE-KMS",
-                "SseAwsKmsKeyId": "arn:aws:kms:us-east-1:644160558196:key/d12e743c-a297-447c-9c13-8f3c08f4161c"
-            },
-            "ConnectionPasswordEncryption": {
-                "ReturnConnectionPasswordEncrypted": true,
-                "AwsKmsKeyId": "arn:aws:kms:us-east-1:644160558196:key/36812ccf-daaf-49b1-a24b-0eef254ffe41"
-            }
-        },
-        "ResponseMetadata": {
-            "RequestId": "0089c68a-ace8-11e9-980b-8b18b4360bda",
-            "HTTPStatusCode": 200,
-            "HTTPHeaders": {
-                "date": "Tue, 23 Jul 2019 01:20:05 GMT",
-                "content-type": "application/x-amz-json-1.1",
-                "content-length": "351",
-                "connection": "keep-alive",
-                "x-amzn-requestid": "0089c68a-ace8-11e9-980b-8b18b4360bda"
-            },
-            "RetryAttempts": 0
-        }
-    }
-}
