--- tests/data/placebo/test_account_glue_encyption_filter/iam.ListAccountAliases_1.json
@@ -1,20 +1,0 @@
-{
-    "status_code": 200,
-    "data": {
-        "AccountAliases": [
-            "custodian-skunk-works"
-        ],
-        "IsTruncated": false,
-        "ResponseMetadata": {
-            "RequestId": "00582ecf-ace8-11e9-a304-0de3d43df0f5",
-            "HTTPStatusCode": 200,
-            "HTTPHeaders": {
-                "x-amzn-requestid": "00582ecf-ace8-11e9-a304-0de3d43df0f5",
-                "content-type": "text/xml",
-                "content-length": "400",
-                "date": "Tue, 23 Jul 2019 01:20:04 GMT"
-            },
-            "RetryAttempts": 0
-        }
-    }
-}
