--- tests/data/placebo/test_account_glue_encyption_filter/tagging.GetResources_1.json
@@ -1,46 +1,0 @@
-{
-    "status_code": 200,
-    "data": {
-        "PaginationToken": "",
-        "ResourceTagMappingList": [
-            {
-                "ResourceARN": "arn:aws:kms:us-east-1:644160558196:key/3f2ab3af-cc4e-427b-b96d-d4356f9764c4",
-                "Tags": [
-                    {
-                        "Key": "Test",
-                        "Value": "Test"
-                    }
-                ]
-            },
-            {
-                "ResourceARN": "arn:aws:kms:us-east-1:644160558196:key/75d1e7ba-f9c5-46b9-97bc-28e6c834aa49",
-                "Tags": [
-                    {
-                        "Key": "Name",
-                        "Value": "c7n/sagemaker"
-                    }
-                ]
-            },
-            {
-                "ResourceARN": "arn:aws:kms:us-east-1:644160558196:key/a476e533-b970-4c74-a341-707420ecf59b",
-                "Tags": [
-                    {
-                        "Key": "Name",
-                        "Value": "CMK-Rotation-Test"
-                    }
-                ]
-            }
-        ],
-        "ResponseMetadata": {
-            "RequestId": "e5d8570c-8867-45f4-ba9c-a788672512a7",
-            "HTTPStatusCode": 200,
-            "HTTPHeaders": {
-                "x-amzn-requestid": "e5d8570c-8867-45f4-ba9c-a788672512a7",
-                "content-type": "application/x-amz-json-1.1",
-                "content-length": "470",
-                "date": "Tue, 23 Jul 2019 01:20:06 GMT"
-            },
-            "RetryAttempts": 0
-        }
-    }
-}
