+++ tests/data/placebo/test_glue_datacat_put_encryption/glue.PutDataCatalogEncryptionSettings_1.json
@@ -1,6 +1,6 @@
 {
-  "status_code": 200,
-  "data": {
-    "ResponseMetadata": {}
-  }
+    "status_code": 200,
+    "data": {
+        "ResponseMetadata": {}
+    }
 }
