--- tests/data/placebo/test_glue_datacat_put_encryption/iam.ListAccountAliases_1.json
@@ -1,10 +1,0 @@
-{
-  "status_code": 200,
-  "data": {
-    "AccountAliases": [
-      "custodian-skunk-works"
-    ],
-    "IsTruncated": false,
-    "ResponseMetadata": {}
-  }
-}
